fixed some warnings
Former-commit-id: a315243db250667fb8f34c7844e08fd721dd1763

+++ src/animation.h
@@ -13,7 +13,7 @@ struct animation
     real32 TimePerFrame;
     uint32 TextureHandle;
     uint32 FrameIndex;
-    real32 CurrentTime;
+    real64 CurrentTime;
     uint32 FrameCount;
     uint32 Rows;
     uint32 Columns;
